requirements.txt

pyyaml
re
time
unittest
